requirements.txt
page.fetch
streamlit
pandas
ast
unicodedata
re
csv
nltk
base64
string
sklearn.metrics
matplotlib.pyplot
BeautifulSoup
urlopen
word_tokenize
PorterStemmer
display_pie_chart